[HIL] Reintroduced temporary handling for firing sequence state
As of this commit the last sample is always used regardless fo the timestamp

+++ src/Motor/HIL/HIL.h
@@ -63,7 +63,7 @@ private:
         const Boardcore::Event& e,
         std::vector<MotorFlightPhases>& changed_flags) override;
 
-    void handleHSMTransition(const uint8_t hsmState);
+    void handleHSMTransition(const float hsmState);
 
     FiringSequenceState previousState = {FiringSequenceState::INIT};
 };
